Integrated Pagination component with ProductGrid component
